+++ kite/src/query/term_scorer.rs
@@ -4,12 +4,12 @@ use similarity::SimilarityModel;
 #[derive(Debug, Clone, PartialEq)]
 pub struct TermScorer {
     pub similarity_model: SimilarityModel,
-    pub boost: f64,
+    pub boost: f32,
 }
 
 
 impl TermScorer {
-    pub fn default_with_boost(boost: f64) -> TermScorer {
+    pub fn default_with_boost(boost: f32) -> TermScorer {
         TermScorer {
             similarity_model: SimilarityModel::Bm25 {
                 k1: 1.2,
@@ -23,6 +23,6 @@ impl TermScorer {
 
 impl Default for TermScorer {
     fn default() -> TermScorer {
-        TermScorer::default_with_boost(1.0f64)
+        TermScorer::default_with_boost(1.0f32)
     }
 }
